Extract benchmark adapter modules
