+++ goal_harness/benchmark_adapters/__init__.py
@@ -1,0 +1,1 @@
+"""Benchmark-specific adapters kept out of the benchmark monolith."""
